collect lingering arithmetic tests
